requirements.txt
openai
langchain
langchain-community        # Extends LangChain with community-built tools
langchain-core             # Core components for building LLM apps
langchain_openai
langchain_experimental
langchainhub
pinecone-client
python-dotenv
tiktoken
docx2txt
pypdf
requests
numpy
pandas
streamlit
scholarly 
autogen 
fastapi 
langchain_groq             # Integrates GROQ features with LangChain
transformers               # to calculate get_token_ids.


